:new: Add bigint global to eslint config

+++ src/unserialize.ts
@@ -1,5 +1,4 @@
 // eslint-disable-next-line import/no-cycle
-/* global BigInt */
 import Parser from './parser'
 import { isInteger, getClass, getIncompleteClass, __PHP_Incomplete_Class, invariant } from './helpers'
 
@@ -49,7 +48,7 @@ function unserializeItem(parser: Parser, scope: Record<string, any>, options: Op
     const value = parser.readUntil(';')
     let parsedValue: number | BigInt = type === 'int' ? parseInt(value, 10) : parseFloat(value)
     if (parsedValue.toString() !== value) {
-      parsedValue = BigInt(value) as BigInt
+      parsedValue = BigInt(value)
     }
     return parsedValue
   }
